+++ GooglePayment/src/main/java/com/braintreepayments/api/GooglePayment.java
@@ -5,6 +5,10 @@ import android.content.Intent;
 import android.content.pm.ActivityInfo;
 import android.text.TextUtils;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
 import com.braintreepayments.api.exceptions.BraintreeException;
 import com.braintreepayments.api.exceptions.ErrorWithResponse;
 import com.braintreepayments.api.exceptions.GoogleApiClientException;
@@ -23,6 +27,7 @@ import com.braintreepayments.api.models.GooglePaymentConfiguration;
 import com.braintreepayments.api.models.GooglePaymentRequest;
 import com.braintreepayments.api.models.MetadataBuilder;
 import com.braintreepayments.api.models.PaymentMethodNonceFactory;
+import com.braintreepayments.api.models.ReadyForGooglePaymentRequest;
 import com.braintreepayments.api.models.TokenizationKey;
 import com.google.android.gms.common.api.ApiException;
 import com.google.android.gms.tasks.OnCompleteListener;
@@ -43,9 +48,6 @@ import org.json.JSONObject;
 
 import java.util.ArrayList;
 import java.util.Collection;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
 
 import static com.braintreepayments.api.GooglePaymentActivity.EXTRA_ENVIRONMENT;
 import static com.braintreepayments.api.GooglePaymentActivity.EXTRA_PAYMENT_DATA_REQUEST;
@@ -77,6 +79,23 @@ public class GooglePayment {
      */
     public static void isReadyToPay(final BraintreeFragment fragment,
                                     final BraintreeResponseListener<Boolean> listener) {
+        isReadyToPay(fragment, null, listener);
+    }
+
+    /**
+     * Before starting the Google Payments flow, use this method to check whether the
+     * Google Payment API is supported and set up on the device. When the listener is called with
+     * {@code true}, show the Google Payments button. When it is called with {@code false}, display other
+     * checkout options.
+     *
+     * @param fragment {@link BraintreeFragment}
+     * @param request {@link ReadyForGooglePaymentRequest}
+     * @param listener Instance of {@link BraintreeResponseListener<Boolean>} to receive the
+     *                 isReadyToPay response.
+     */
+    public static void isReadyToPay(final BraintreeFragment fragment,
+                                    final @Nullable ReadyForGooglePaymentRequest request,
+                                    final BraintreeResponseListener<Boolean> listener) {
         try {
             Class.forName(PaymentsClient.class.getName());
         } catch (ClassNotFoundException | NoClassDefFoundError e) {
@@ -101,7 +120,6 @@ public class GooglePayment {
                                 .setEnvironment(getEnvironment(configuration.getGooglePayment()))
                                 .build());
 
-
                 JSONObject json = new JSONObject();
                 JSONArray allowedCardNetworks = new JSONArray();
 
@@ -121,6 +139,11 @@ public class GooglePayment {
                                                             .put("PAN_ONLY")
                                                             .put("CRYPTOGRAM_3DS"))
                                                     .put("allowedCardNetworks", allowedCardNetworks))));
+
+                    if (request != null) {
+                        json.put("existingPaymentMethodRequired", request.isExistingPaymentMethodRequired());
+                    }
+
                 } catch (JSONException ignored) {
                 }