+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Braintree Android Google Pay SDK Release Notes
 
+## unreleased
+* Add support for Google Pay's `existingPaymentMethodRequired` option
+
 ## 3.2.0
 * Add support for `isNetworkTokenized`